Extend nemo.utils package with new tools
